Implement the pipeline experiment with averaged contextualized embeddings
The previous pipeline experiment used the "empty template" to obtain a unique representation of words. However, after several experiments, the word embedding seems to be influenced more from the original word than from the context (i.e. the "template"). Thus, we substitute the empty template with multiple non-empty templates. The final word embedding is the average of all the embeddings obtained for the word from multiple templates.